icp/pyproject.toml
[tool.poetry]
name = "icp"
version = "0.1.0"
description = ""
authors = ["HideBa <baba.papa1120.ba@gmail.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.23"
scikit-learn = "^1.5.2"
pyyaml = "^6.0.2"
addict = "^2.4.0"
open3d = "^0.18.0"
tqdm = "^4.66.5"
pillow = "^10.4.0"
pandas = "^2.2.2"
mypy = "^1.11.2"
laspy = "^2.5.4"
pycpd = "^2.0.0"
rerun-sdk = "^0.18.2"
matplotlib = "^3.9.2"
simpleicp = "^2.0.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
